requirements.txt
feedparser
google-generativeai
python-wordpress-xmlrpc
requests